// frontend/homework/18/src/interfaces/Interface.ts
// Generated by https://quicktype.io

export interface IAPIResponse {
    id:          number;
    title:       string;
    price:       number;
    description: string;
    category:    string;
    image:       string;
    rating:      Rating;
}

export interface Rating {
    rate:  number;
    count: number;
}